refactor sidebar search overlay
